refactor: move TestCounts + TestScopeOutput to homeboy-extension-contract
Continues the test phase-subsystem breakdown. TestCounts (pure 4-u64 struct +
pure new() constructor) and TestScopeOutput (pure changed-scope selection
result) move to the contract crate. The behavior around them (baseline
comparison fingerprinting, changed-scope computation functions) stays in core.

Core re-exports both so all crate::extension::test paths stay stable. Clears the
TestCounts + TestScopeOutput core-import edges. The aggregate TestCommandOutput/
LintCommandOutput remain in core - they are top-level result clusters (like
BenchResults was) still blocked by field types in report/run/workflow modules.
